requirements.txt
ag-ui-protocol==0.1.8
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.67.0
anyio==4.10.0
appnope==0.1.4
argcomplete==3.6.2
asttokens==3.0.0
attrs==25.3.0
boto3==1.40.2
botocore==1.40.2
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
cohere==5.18.0
colorama==0.4.6
comm==0.2.3
debugpy==1.8.15
decorator==5.2.1
distro==1.9.0
eval_type_backport==0.2.2
executing==2.2.0
fastapi==0.116.1
fastavro==1.12.0
fastuuid==0.12.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
genai-prices==0.0.27
gitdb==4.0.12
GitPython==3.1.40
google-auth==2.40.3
google-genai>=1.36.0
googleapis-common-protos==1.70.0
griffe==1.9.0
groq==0.30.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.3
keyring>=24.0.0
idna==3.10
importlib_metadata==8.7.0
invoke==2.2.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
litellm==1.77.0
logfire==4.1.0
logfire-api==4.1.0
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mcp==1.12.3
mdurl==0.1.2
mermaid-parser-py==0.0.2
mermaid-py==0.8.0
mistralai==1.9.10
multidict==6.6.3
nest-asyncio==1.6.0
networkx==3.5
nexus-rpc==1.1.0
openai==1.107.1
openai-agents==0.3.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-httpx==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.8
pminit==1.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycares==4.11.0
pycparser==2.23
pydantic==2.11.7
pydantic-ai==1.0.6
pydantic-ai-slim==1.0.6
pydantic-evals==1.0.6
pydantic-graph==1.0.6
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pythonmonkey==1.2.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.47.2
temporalio==1.17.0
tenacity==8.5.0
tiktoken==0.10.0
tokenizers==0.21.4
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
tree-sitter==0.23.2
tree-sitter-c==0.21.4
tree-sitter-c-sharp==0.23.1
tree-sitter-cpp==0.23.4
tree-sitter-embedded-template==0.23.2
tree-sitter-java==0.23.5
tree-sitter-javascript==0.21.4
tree-sitter-language-pack==0.8.0
tree-sitter-python==0.23.6
tree-sitter-typescript==0.21.2
tree-sitter-yaml==0.7.1
types-protobuf==6.30.2.20250822
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
